Week2/requirements.txt
chromadb
sentence-transformers
pandas
openai
python-dotenv